Script: Ensure that functions are called within valid container
This adds a new argument to call() - container. The underlying type may
check if the container passed to function is valid for the function (it
won't be so if function was copied to another, non-compatible object).

This commit makes NativeFunction a template (to allow automatic checking
of container) and moves it to separate file.