Script: Remove one layer of indirection in scopes
This makes ScopeObjects stored as values on call stack. Also,
stronger handling of 'this' object resolution and containers was
introduced to avoid bugs.

Other things that was changed:
- Added fancy exceptions when trying to access 'this' in invalid
conditions
- Removed "<global scope>" as it would need to store pointer to
global object to work properly. Statements not wrapped inside
function are run on empty stack, which is pushed to when creating
a block statement.

This improves "primes" test runtime to ~420 ms :)